+++ Sources/MapboxMaps/Style/MapStyleReconciler.swift
@@ -1,78 +1,97 @@
-/// Incrementally applies changes in map style.
-class MapStyleReconciler {
+import os
+
+/// Style reconciler incrementally applies style changes to the map.
+/// Use `mapStyle` setter or `loadStyle` function to update style or import configurations.
+/// The update will trigger actual style load only if needed.
+final class MapStyleReconciler {
     var mapStyle: MapStyle? {
-        didSet {
-            reconcile(from: oldValue)
-        }
-    }
-
-    struct LoadingState {
-        var oldImports: [StyleImportConfiguration]?
-        var token: AnyCancelable?
-    }
-
-    private let onStyleDataLoaded: Signal<StyleDataLoaded>
-    private let coreStyleManager: StyleManagerProtocol
-    private var loadingState: LoadingState?
-
-    init(
-        coreStyleManager: StyleManagerProtocol,
-        onStyleDataLoaded: Signal<StyleDataLoaded>
-    ) {
-        self.coreStyleManager = coreStyleManager
-        self.onStyleDataLoaded = onStyleDataLoaded
-    }
-
-    private func reconcile(from old: MapStyle?) {
-        if let mapStyle, old?.loadMethod != mapStyle.loadMethod {
-            switch mapStyle.loadMethod {
-            case let .json(json):
-                coreStyleManager.setStyleJSONForJson(json)
-            case let .uri(uri):
-                coreStyleManager.setStyleURIForUri(uri.rawValue)
+        get { _mapStyle }
+        set {
+            if let newValue {
+                loadStyle(newValue)
+            } else {
+                _mapStyle = nil
             }
         }
-
-        switch (loadingState, coreStyleManager.isStyleLoaded()) {
-        case (.none, true):
-            // Style is loaded, just need update the import config
-            reconcileStyleImports(from: old?.importConfigurations)
-        case (.none, false):
-            // Style began loading, need to wait loading to update import config
-            let token = onStyleDataLoaded
-                .filter { $0.type == .style }
-                .observeNext { [weak self] _ in
-                    self?.handleStyleDataLoaded()
-                }
-            loadingState = LoadingState(oldImports: old?.importConfigurations, token: token)
-        case (.some(_), _):
-            // Already waiting style to load. When style loads new configuration will be applied.
-            break
-        }
     }
 
-    func handleStyleDataLoaded() {
-        guard let loadingState else {
-            assertionFailure()
+    private let styleManager: StyleManagerProtocol
+    private var pendingCompletions = [(Error?) -> Void]()
+    private var _mapStyle: MapStyle?
+
+    init(styleManager: StyleManagerProtocol) {
+        self.styleManager = styleManager
+    }
+
+    func loadStyle(
+        _ style: MapStyle,
+        transition: TransitionOptions? = nil,
+        completion: ((Error?) -> Void)? = nil
+    ) {
+        if _mapStyle?.loadMethod != style.loadMethod {
+            _mapStyle = style
+            let callbacks = RuntimeStylingCallbacks(
+                layers: { [weak self] in
+                    if let transition {
+                        self?.styleManager.setStyleTransitionFor(transition)
+                    }
+                    // When new style is loaded, no need in incremental change of import configuration.
+                    self?.reconcileStyleImports(from: nil)
+                },
+                completed: { [weak self] in
+                    completion?(nil)
+                    self?.completeLoad(nil)
+                },
+                cancelled: { [weak self] in
+                    let error = CancelError()
+                    completion?(error)
+                    self?.completeLoad(error)
+                },
+                error: { [weak self] error in
+                    completion?(error)
+                    self?.completeLoad(error)
+                })
+
+            switch style.loadMethod {
+            case let .json(json):
+                styleManager.setStyleJSON(json, callbacks: callbacks)
+            case let .uri(uri):
+                styleManager.setStyleURI(uri.rawValue, callbacks: callbacks)
+            }
             return
         }
+        let old = _mapStyle
+        _mapStyle = style
 
-        reconcileStyleImports(from: loadingState.oldImports)
-        self.loadingState = nil
+        if styleManager.isStyleLoaded() {
+            reconcileStyleImports(from: old?.importConfigurations)
+            completion?(nil)
+        } else {
+            // The style with the same uri is already loading, save completion for future execution.
+            pendingCompletions += completion.asArray
+        }
     }
 
-    func reconcileStyleImports(from old: [StyleImportConfiguration]?) {
+    private func completeLoad(_ error: Error?) {
+        let completions = pendingCompletions
+        pendingCompletions.removeAll()
+        for completion in completions {
+            completion(error)
+        }
+    }
+
+    private func reconcileStyleImports(from old: [StyleImportConfiguration]?) {
         guard let mapStyle else { return }
         Self.reconcileStyleImports(
             from: old,
             to: mapStyle.importConfigurations,
-            coreStyleManager: coreStyleManager)
+            styleManager: styleManager)
     }
 
     static func reconcileStyleImports(
         from old: [StyleImportConfiguration]?,
         to new: [StyleImportConfiguration],
-        coreStyleManager: StyleManagerProtocol
+        styleManager: StyleManagerProtocol
     ) {
         for importConfig in new {
             let oldImportConfig = old?.first { $0.importId == importConfig.importId }
@@ -82,7 +101,7 @@ class MapStyleReconciler {
                         continue
                     }
                     try handleExpected {
-                        coreStyleManager.setStyleImportConfigPropertyForImportId(importConfig.importId, config: key, value: value.rawValue)
+                        styleManager.setStyleImportConfigPropertyForImportId(importConfig.importId, config: key, value: value.rawValue)
                     }
                 }
             } catch {